+++ android/src/main/java/com/getcapacitor/community/tts/TextToSpeech.java
@@ -13,6 +13,7 @@ import android.util.Log;
 import com.getcapacitor.JSArray;
 import com.getcapacitor.JSObject;
 import java.util.ArrayList;
+import java.util.Comparator;
 import java.util.HashMap;
 import java.util.Locale;
 import java.util.Set;
@@ -52,6 +53,7 @@ public class TextToSpeech implements android.speech.tts.TextToSpeech.OnInitListe
         float rate,
         float pitch,
         float volume,
+        int voice,
         String callbackId,
         SpeakResultCallback resultCallback
     ) {
@@ -83,6 +85,15 @@ public class TextToSpeech implements android.speech.tts.TextToSpeech.OnInitListe
             tts.setLanguage(locale);
             tts.setSpeechRate(rate);
             tts.setPitch(pitch);
+
+            if (voice >= 0) {
+                ArrayList<Voice> supportedVoices = getSupportedVoicesOrdered();
+                if (voice < supportedVoices.size()) {
+                    Voice newVoice = supportedVoices.get(voice);
+                    int resultCode = tts.setVoice(newVoice);
+                }
+            }
+
             tts.speak(text, android.speech.tts.TextToSpeech.QUEUE_FLUSH, ttsParams, callbackId);
         } else {
             HashMap<String, String> ttsParams = new HashMap<>();
@@ -111,9 +122,22 @@ public class TextToSpeech implements android.speech.tts.TextToSpeech.OnInitListe
         return result;
     }
 
+    public ArrayList<Voice> getSupportedVoicesOrdered() {
+        Set<Voice> supportedVoices = tts.getVoices();
+        ArrayList<Voice> orderedVoices = new ArrayList<Voice>();
+        for (Voice supportedVoice : supportedVoices) {
+            orderedVoices.add(supportedVoice);
+        }
+
+        Comparator<Voice> voiceComparator = Comparator.comparing(v -> v.hashCode());
+        orderedVoices.sort(voiceComparator);
+
+        return orderedVoices;
+    }
+
     public JSArray getSupportedVoices() {
         ArrayList<JSObject> voices = new ArrayList<>();
-        Set<Voice> supportedVoices = tts.getVoices();
+        ArrayList<Voice> supportedVoices = getSupportedVoicesOrdered();
         for (Voice supportedVoice : supportedVoices) {
             JSObject obj = this.convertVoiceToJSObject(supportedVoice);
             voices.add(obj);