+++ ios/Plugin/TextToSpeechPlugin.swift
@@ -1,5 +1,6 @@
 import Foundation
 import Capacitor
+import AVFoundation
 
 /**
  * Please read the Capacitor iOS Plugin Development Guide
@@ -17,6 +18,7 @@ public class TextToSpeechPlugin: CAPPlugin {
         let rate = call.getFloat("rate") ?? 1.0
         let pitch = call.getFloat("pitch") ?? 1.0
         let volume = call.getFloat("volume") ?? 1.0
+        let voice = call.getInt("voice") ?? -1
         let category = call.getString("category") ?? "ambient"
 
         let isLanguageSupported = implementation.isLanguageSupported(lang)
@@ -26,7 +28,7 @@ public class TextToSpeechPlugin: CAPPlugin {
         }
 
         do {
-            try implementation.speak(text, lang, rate, pitch, category, volume, call)
+            try implementation.speak(text, lang, rate, pitch, category, volume, voice, call)
         } catch {
             call.reject(error.localizedDescription)
         }
@@ -49,8 +51,22 @@ public class TextToSpeechPlugin: CAPPlugin {
     }
 
     @objc func getSupportedVoices(_ call: CAPPluginCall) {
+        let allVoices = AVSpeechSynthesisVoice.speechVoices()
+        var res: [[String: Any]] = []
+
+        for voice in allVoices {
+            let lang = [
+                "default": false,
+                "lang": voice.language,
+                "localService": true,
+                "name": voice.name,
+                "voiceURI": voice.identifier
+            ] as [String: Any]
+            res.append(lang)
+        }
+
         call.resolve([
-            "voices": []
+            "voices": res
         ])
     }